Moved RequestContextTests into test_context.
